🚨 split TS and Flow unit tests
